requirements.txt
python-telegram-bot[webhooks]==20.7
python-dotenv==1.0.0
openai>=1.0.0
